requirements.txt
##
## Runtime dependencies for scripts in this repository
##
## Notes for maintainers:
## - Prefer pinning with lower bounds only (>=) to allow patch/minor updates in CI.
## - Keep versions compatible with the GitHub Actions runners used in workflows.
## - These are intentionally lightweight; project‑specific deps live in service repos.
##
PyYAML>=6.0
jsonschema>=4.17.0
Jinja2>=3.1.0
Click>=8.1.0
pytest>=7.2.0
requests>=2.31.0
python-dateutil>=2.8.0
packaging>=23.0
matplotlib>=3.7.0
plotly>=5.14.0
urllib3>=2.0.0
